+++ Changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 **latest**
+- Configure pydio temporary directory
 
 **8.0.1-1**
 - Fix vhosts to catch no existing files, important to support shared files